+++ tests/Nexendrie/Rss/Extensions/DublinCoreTest.php
@@ -3,6 +3,7 @@ declare(strict_types=1);
 
 namespace Nexendrie\Rss\Extensions;
 
+use DateTime;
 use Nexendrie\Rss\Collection;
 use Nexendrie\Rss\Generator;
 use Nexendrie\Rss\RssChannelItem;
@@ -37,7 +38,7 @@ final class DublinCoreTest extends \Tester\TestCase
         $generator->dataSource = static function () use ($extensionName, $elementName1, $elementName2) {
             $collection = new Collection();
             $collection[] = new RssChannelItem([
-                "title" => "Item 1", "description" => "Item 1 description", "link" => "", "pubDate" => 123,
+                "title" => "Item 1", "description" => "Item 1 description", "link" => "", "pubDate" => new DateTime(),
                 "$extensionName:$elementName1" => "abc", "$extensionName:$elementName2" => "def",
             ]);
             return $collection;