+++ tests/Nexendrie/Rss/Extensions/SlashTest.php
@@ -3,6 +3,7 @@ declare(strict_types=1);
 
 namespace Nexendrie\Rss\Extensions;
 
+use DateTime;
 use Nexendrie\Rss\Collection;
 use Nexendrie\Rss\Generator;
 use Nexendrie\Rss\RssChannelItem;
@@ -39,7 +40,7 @@ final class SlashTest extends \Tester\TestCase
         $generator->dataSource = static function () use ($extensionName, $elementName1, $elementName2, $elementName3, $elementName4) {
             $collection = new Collection();
             $collection[] = new RssChannelItem([
-                "title" => "Item 1", "description" => "Item 1 description", "link" => "", "pubDate" => 123,
+                "title" => "Item 1", "description" => "Item 1 description", "link" => "", "pubDate" => new DateTime(),
                 "$extensionName:$elementName1" => "abc", "$extensionName:$elementName2" => "def",
                 "$extensionName:$elementName3" => 1, "$extensionName:$elementName4" => "1,2,3",
             ]);